+++ sshd-sftp/src/main/java/org/apache/sshd/server/subsystem/sftp/SftpFileSystemAccessor.java
@@ -73,6 +73,8 @@ public interface SftpFileSystemAccessor {
      *
      * @param session The {@link ServerSession} through which the request was received
      * @param subsystem The SFTP subsystem instance that manages the session
+     * @param fileHandle The {@link FileHandle} representing the created channel - may be
+     * {@code null} if not invoked within the context of such a handle (special cases)
      * @param file The requested <U>local</U> file {@link Path}
      * @param handle The assigned file handle through which the remote peer references this file.
      * May be {@code null}/empty if the request is due to some internal functionality
@@ -83,7 +85,7 @@ public interface SftpFileSystemAccessor {
      * @throws IOException If failed to open
      */
     default SeekableByteChannel openFile(
-            ServerSession session, SftpEventListenerManager subsystem,
+            ServerSession session, SftpEventListenerManager subsystem, FileHandle fileHandle,
             Path file, String handle, Set<? extends OpenOption> options, FileAttribute<?>... attrs)
                 throws IOException {
         /*
@@ -103,9 +105,11 @@ public interface SftpFileSystemAccessor {
      *
      * @param session The {@link ServerSession} through which the request was received
      * @param subsystem The SFTP subsystem instance that manages the session
+     * @param fileHandle The {@link FileHandle} representing the created channel
      * @param file The requested <U>local</U> file {@link Path}
      * @param handle The assigned file handle through which the remote peer references this file
-     * @param channel The original {@link Channel} that was returned by {@link #openFile(ServerSession, SftpEventListenerManager, Path, String, Set, FileAttribute...)}
+     * @param channel The original {@link Channel} that was returned by
+     * {@link #openFile(ServerSession, SftpEventListenerManager, FileHandle, Path, String, Set, FileAttribute...)}
      * @param position The position at which the locked region is to start - must be non-negative
      * @param size The size of the locked region; must be non-negative, and the sum
      * <tt>position</tt>&nbsp;+&nbsp;<tt>size</tt> must be non-negative
@@ -115,7 +119,9 @@ public interface SftpFileSystemAccessor {
      * @throws IOException If failed to honor the request
      * @see FileChannel#tryLock(long, long, boolean)
      */
-    default FileLock tryLock(ServerSession session, SftpEventListenerManager subsystem,
+    @SuppressWarnings("checkstyle:ParameterNumber")
+    default FileLock tryLock(
+            ServerSession session, SftpEventListenerManager subsystem, FileHandle fileHandle,
             Path file, String handle, Channel channel, long position, long size, boolean shared)
                 throws IOException {
         if (!(channel instanceof FileChannel)) {
@@ -130,15 +136,17 @@ public interface SftpFileSystemAccessor {
      *
      * @param session The {@link ServerSession} through which the request was received
      * @param subsystem The SFTP subsystem instance that manages the session
+     * @param fileHandle The {@link FileHandle} representing the created channel
      * @param file The requested <U>local</U> file {@link Path}
      * @param handle The assigned file handle through which the remote peer references this file
-     * @param channel The original {@link Channel} that was returned by {@link #openFile(ServerSession, SftpEventListenerManager, Path, String, Set, FileAttribute...)}
+     * @param channel The original {@link Channel} that was returned by
+     * {@link #openFile(ServerSession, SftpEventListenerManager, FileHandle, Path, String, Set, FileAttribute...)}
      * @throws IOException If failed to execute the request
      * @see FileChannel#force(boolean)
      * @see <A HREF="https://github.com/openssh/openssh-portable/blob/master/PROTOCOL">OpenSSH -  section 10</A>
      */
     default void syncFileData(
-            ServerSession session, SftpEventListenerManager subsystem, Path file, String handle, Channel channel)
+            ServerSession session, SftpEventListenerManager subsystem, FileHandle fileHandle, Path file, String handle, Channel channel)
                 throws IOException {
         if (!(channel instanceof FileChannel)) {
             throw new StreamCorruptedException("Non file channel to sync: " + channel);
@@ -148,18 +156,67 @@ public interface SftpFileSystemAccessor {
     }
 
     /**
+     * Called to inform the accessor that it should close the file
+     *
+     * @param session The {@link ServerSession} through which the request was received
+     * @param subsystem The SFTP subsystem instance that manages the session
+     * @param fileHandle The {@link FileHandle} representing the created channel - may be
+     * {@code null} if not invoked within the context of such a handle (special cases)
+     * @param file The requested <U>local</U> file {@link Path}
+     * @param handle The assigned file handle through which the remote peer references this file
+     * @param channel The original {@link Channel} that was returned by
+     * {@link #openFile(ServerSession, SftpEventListenerManager, FileHandle, Path, String, Set, FileAttribute...)}
+     * @param options The original options used to open the channel
+     * @throws IOException If failed to execute the request
+     */
+    default void closeFile(
+            ServerSession session, SftpEventListenerManager subsystem, FileHandle fileHandle,
+            Path file, String handle, Channel channel, Set<? extends OpenOption> options)
+                throws IOException {
+        if ((channel == null) || (!channel.isOpen())) {
+            return;
+        }
+
+        channel.close();
+    }
+
+    /**
      * Called when a new directory stream is requested
      *
      * @param session The {@link ServerSession} through which the request was received
      * @param subsystem The SFTP subsystem instance that manages the session
+     * @param dirHandle The {@link DirectoryHandle} representing the stream
      * @param dir The requested <U>local</U> directory
      * @param handle The assigned directory handle through which the remote peer references this directory
      * @return The opened {@link DirectoryStream}
      * @throws IOException If failed to open
      */
     default DirectoryStream<Path> openDirectory(
-            ServerSession session, SftpEventListenerManager subsystem, Path dir, String handle)
+            ServerSession session, SftpEventListenerManager subsystem, DirectoryHandle dirHandle, Path dir, String handle)
                 throws IOException {
         return Files.newDirectoryStream(dir);
+    }
+
+    /**
+     * Called when a directory stream is no longer required
+     *
+     * @param session The {@link ServerSession} through which the request was received
+     * @param subsystem The SFTP subsystem instance that manages the session
+     * @param dirHandle The {@link DirectoryHandle} representing the stream - may be
+     * {@code null} if not invoked within the context of such a handle (special cases)
+     * @param dir The requested <U>local</U> directory
+     * @param handle The assigned directory handle through which the remote peer references this directory
+     * @param ds The disposed {@link DirectoryStream}
+     * @throws IOException If failed to open
+     */
+    default void closeDirectory(
+            ServerSession session, SftpEventListenerManager subsystem, DirectoryHandle dirHandle,
+            Path dir, String handle, DirectoryStream<Path> ds)
+                throws IOException {
+        if (ds == null) {
+            return; // debug breakpoint
+        }
+
+        ds.close();
     }
 }